Set the 404 page main header height to 102px
On normal article pages without featured images, there is a cite
element that helps to extend the height of the main header. On the
404 pages, there is no cite element, so we need to manually
override things a bit.

In the future it may make more sense to define the main header at
a specific height everywhere rather than using auto